agent: refine host upgrade preflight
